fixed bug with loading spinner, sent poll with history
